requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.10.3

faster-whisper==1.0.3
requests==2.32.3
huggingface_hub==0.26.2

sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
reportLab